[dm-bufio] Factor out an lru abstraction.
Merely pulls together a linked list and count atm, but we're going to
switch to a clock alg later.

Methods are provided to evict the oldest entry that passes a predicate, and
to iterate through the entries.

Debug code also present for sanity checking the lru.  This is trivial
atm since it's just checking the linked list implementation which we
have confidence in.  But when we switch to clock alg it will be more
useful.